whisper/requirements.txt
openai-whisper
tiktoken
ffmpeg-python